# docs/openapi/royalgpt-command-api.yaml
openapi: 3.1.0
...
info:
  title: RoyalGPT Command API
  version: 2.1.0
  description: |
    Royal Equips orchestrator surface for RoyalGPT-enabled clients.
    Provides production-ready health diagnostics, product intelligence,
    fraud detection, and operational analytics endpoints consumed by
    RoyalGPT agents.
servers:
  - url: https://command.royalequips.nl
    description: Production deployment
  - url: http://localhost:10000
    description: Local development
  - url: https://staging.command.royalequips.nl
    description: Staging environment
...
paths:
...
  /fraud/scan:
    post:
      summary: Execute an on-demand fraud scan
      description: |
        Triggers the fraud detection pipeline managed by the security agent.
        Applies machine-learning heuristics, rule-based checks, and
        orchestration fallbacks. Requires administrative authorization and
        is rate limited in parity with `/api/security/fraud-detection/run`.
      tags:
        - Fraud
      responses:
        '200':
          description: Fraud scan completed successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FraudScanResult'
        '401':
          description: Authentication required.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Rate limit exceeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Fraud detection agent unavailable.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
...
    FraudScanResult:
      type: object
      required:
        - status
        - summary
        - metrics
        - alerts
        - timestamp
      properties:
        status:
          type: string
          enum: [completed, partial, failed]
        summary:
          type: object
          required:
            - analyzed
            - highRisk
            - alertsGenerated
          properties:
            analyzed:
              type: integer
              minimum: 0
            highRisk:
              type: integer
              minimum: 0
            alertsGenerated:
              type: integer
              minimum: 0
            durationMs:
              type: integer
              minimum: 0
        metrics:
          type: object
          properties:
            averageRiskScore:
              type: number
            threshold:
              type: number
            falsePositiveRate:
              type: number
        alerts:
          type: array
          items:
            type: object
            required:
              - id
              - type
              - riskScore
              - action
            properties:
              id:
                type: string
              type:
                type: string
              riskScore:
                type: number
              action:
                type: string
              orderId:
                type: string
              customerId:
                type: string
              reasons:
                type: array
                items:
                  type: string
        metadata:
          type: object
          properties:
            triggeredBy:
              type: string
            correlationId:
              type: string
        timestamp:
          type: string
          format: date-time
...
    ErrorResponse:
      type: object
      required:
        - error
        - message
        - timestamp
      properties:
        error:
          type: string
        message:
          type: string
        timestamp:
          type: string
          format: date-time
        details:
          type: object
          additionalProperties: true